refactor: reorganize prompt tools and move related functions to new module
