+++ ee/tabby-webserver/src/service/utils/mod.rs
@@ -1,0 +1,1 @@
+pub mod prompt;
